[SDPA] Port the SM80 backward onto the SdpaBwdDsl adapter path
SM80 now follows the SM120 backward lowering shape exactly (design doc S5 /
F-3, the backward analogue of #682): one adapter class (SdpaBwdDslSm80 in
api_dsl.py) implementing the SdpaBwdDsl contract, lowered through the shared
lower_dsl_bwd; bwd/api.py is deleted.

- SdpaBwdDslSm80: descriptor-level check_support (flavor pick, mask
  resolution, dense_flex + strided-stats acceptance), no-op compile (the
  kernels self-cache until the TemplateParams conversion), execute with the
  full issue #514 carving (pad/gather staging, strided-stats gather, kernel
  workspace tail) and the d64 fast-path routing. SM80-only operands (bias ->
  dBias, RoPE) are extra optional keywords, as the contract permits.
- lower_dsl_bwd is parameterized by api_type (mirrors lower_dsl_prefill) and
  now drives both backward cells; SM80-only constructor facts and execute
  operands forward via signature introspection, so the SM120 adapter is
  untouched. lower_sm80_bwd (the plan-time APIBase half-way house) is gone.
- sdpa_bwd_wrapper_sm80 keeps its public signature (dense via the adapter,
  packed THD via the kernel varlen path); SdpabwdSm80 is replaced by
  SdpaBwdDslSm80 (experimental API, exports/tests updated).
- The SM80 suites declare their backward output strides explicitly: the
  shared lowering honors DECLARED port geometry (IR-inferred output strides
  are provisional row-major -- the layout invariant), where the old adapter
  leniently trusted caller tensor metadata.

Verified on A100: SM80 suites all levels 118 passed; test_mhas_v2 fwd+bwd
L0 = 421/0 at 100% FROST routing (fwd 623, bwd 176) through the ported path.
